theme: make theme files independent to commonColor
