+++ src/basic/Checkbox.js
@@ -6,7 +6,7 @@ import { connectStyle } from 'native-base-shoutem-theme';
 
 import mapPropsToStyleNames from '../utils/mapPropsToStyleNames';
 import variable from '../theme/variables/platform';
-import { PLATFORM } from '../theme/variables/commonColor';
+import { PLATFORM } from '../theme/variables/common';
 import computeProps from '../utils/computeProps';
 
 class CheckBox extends Component {
