+++ src/basic/ToastContainer.js
@@ -4,7 +4,7 @@ import { Keyboard, Platform, Animated, ViewPropTypes } from 'react-native';
 import { connectStyle } from 'native-base-shoutem-theme';
 
 import mapPropsToStyleNames from '../utils/mapPropsToStyleNames';
-import { PLATFORM } from '../theme/variables/commonColor';
+import { PLATFORM } from '../theme/variables/common';
 
 import { Text } from './Text';
 import { Button } from './Button';
