+++ src/theme/components/CardItem.js
@@ -3,7 +3,7 @@
 import { StyleSheet } from 'react-native';
 
 import variable from './../variables/platform';
-import { PLATFORM } from './../variables/commonColor';
+import { PLATFORM } from './../variables/common';
 
 export default (variables /* : * */ = variable) => {
   const platform = variables.platform;
