+++ src/theme/variables/common.js
@@ -1,0 +1,18 @@
+import { Platform, Dimensions } from 'react-native';
+
+export const PLATFORM = {
+  ANDROID: 'android',
+  IOS: 'ios',
+  MATERIAL: 'material',
+  WEB: 'web'
+};
+
+export const deviceHeight = Dimensions.get('window').height;
+export const deviceWidth = Dimensions.get('window').width;
+
+export const isIphoneX =
+  Platform.OS === PLATFORM.IOS &&
+  (deviceHeight === 812 ||
+    deviceWidth === 812 ||
+    deviceHeight === 896 ||
+    deviceWidth === 896);
