+++ src/theme/variables/material.js
@@ -1,20 +1,12 @@
 // @flow
 
 import color from 'color';
-import { Platform, Dimensions, PixelRatio } from 'react-native';
+import { Platform, PixelRatio } from 'react-native';
 
-import { PLATFORM } from './commonColor';
+import { PLATFORM, deviceHeight, deviceWidth, isIphoneX } from './common';
 
-const deviceHeight = Dimensions.get('window').height;
-const deviceWidth = Dimensions.get('window').width;
 const platform = Platform.OS;
 const platformStyle = PLATFORM.MATERIAL;
-const isIphoneX =
-  platform === PLATFORM.IOS &&
-  (deviceHeight === 812 ||
-    deviceWidth === 812 ||
-    deviceHeight === 896 ||
-    deviceWidth === 896);
 
 export default {
   platformStyle,
